requirements.txt
pandas 
numpy
seaborn
scikit-learn
python-box
pyYAML
mlflow
matplotlib
tqdm
ensure
joblib
types-pyYAML
Flask
Flask-Cors